Backend/chat/requirements.txt
mysql-connector-python>=8.0.0
spacy>=3.7.0
python-dotenv>=1.0.0
